Sec7_Jackson_Network_Sim/requirements.txt
gymnasium==0.29.1
hydra-core==1.3.2
matplotlib==3.8.4
numpy==1.26.4
queueing_tool==1.2.5
ray==2.12.0
pyproximal==0.9.0
dm_tree==0.1.8
pyarrow==17.0.0
pandas==2.2.3
tensorboardX==2.6.2.2
typer==0.12.5
scikit-image==0.24.0
fsspec==2024.9.0